Add runner for local environment using `testcontainers`
